Refactor has_access for re-usability.
